[Refactor] didHandler to didHelper to be more inline with nomenclature used throughout the project. (#47)
