+++ src/utils/didHelper.ts
