+++ test/verifier/verifyPresentationHelper.test.ts
@@ -9,12 +9,12 @@ import stringify from 'fast-json-stable-stringify';
 import { verifyPresentationHelper as verifyPresentation } from '../../src/verifier/verifyPresentationHelper';
 import { CredentialRequest } from '@unumid/types';
 import { JSONObj } from '../../src/types';
-import { getDIDDoc } from '../../src/utils/didHandler';
+import { getDIDDoc } from '../../src/utils/didHelper';
 import { makeNetworkRequest } from '../../src/utils/networkRequestHelper';
 import { doVerify } from '../../src/utils/verify';
 
-jest.mock('../../src/utils/didHandler', () => {
-  const actual = jest.requireActual('../../src/utils/didHandler');
+jest.mock('../../src/utils/didHelper', () => {
+  const actual = jest.requireActual('../../src/utils/didHelper');
   return {
     ...actual,
     getDIDDoc: jest.fn()
